money-page: abstract the  MoneyDialPad
